Finished Assignment
Scrolling marquee was implemented to scroll left until the mouse hovers over. When the mouse hovers over it the scrolling stops and any text the mouse covers underlines and makes the mouse point to give the idea of a link. Overlay text was also implemented so that when the mouse hovers over the image of LeBron James a text box appears at the bottom with his name.